requirements.txt
# Microsoft Fabric Agentic Workloads Lab - Python Dependencies

# Core Azure and Authentication
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.0
azure-core>=1.29.0

# Fabric and Power BI
# Note: fabric-sdk may need to be installed from preview/private feed
# For now, use REST API with requests
requests>=2.31.0

# Azure OpenAI and AI
openai>=1.12.0
azure-ai-openai>=1.0.0b1

# Agent Frameworks
semantic-kernel>=0.9.0
langchain>=0.1.0
langchain-openai>=0.0.5

# Data Processing
pandas>=2.1.0
pyarrow>=14.0.0
numpy>=1.24.0

# Spark (for local development)
pyspark>=3.5.0
delta-spark>=3.0.0

# ML and MLflow
mlflow>=2.10.0
scikit-learn>=1.4.0

# Azure Event Hubs (for real-time)
azure-eventhub>=5.11.0

# Database connectivity
pyodbc>=5.0.0
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
jsonschema>=4.20.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipykernel>=6.29.0

# Monitoring and Logging
opencensus>=0.11.0
opencensus-ext-azure>=1.1.0
applicationinsights>=0.11.0

# HTTP and API
aiohttp>=3.9.0
httpx>=0.26.0

# Cost tracking
tiktoken>=0.6.0  # For token counting
